JWT userId checks added to UserController

+++ src/main/java/jotbackend/controllers/UserController.java
@@ -1,9 +1,13 @@
 package jotbackend.controllers;
 
+import jotbackend.classes.JotUserDetails;
 import jotbackend.classes.User;
 import jotbackend.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import java.util.Date;
@@ -14,6 +18,8 @@ import java.util.Optional;
 public class UserController {
     @Autowired
     private UserRepository userRepository;
+/*
+    // Admin functionality
 
     @PostMapping(path = "/add")
     public @ResponseBody String addNewUser (@RequestParam String firstName, @RequestParam String lastName, @RequestParam String emailAddress,
@@ -30,29 +36,70 @@ public class UserController {
 
     @GetMapping(path = "/all")
     public @ResponseBody Iterable<User> getAllUsers() { return userRepository.findAll();}
+ */
 
-    @GetMapping(path = "/{userId}")
-    public @ResponseBody
-    Optional<User> getUserById(@PathVariable Integer userId) { return userRepository.findById(userId);}
+    @GetMapping(path = "/me")
+    public ResponseEntity getUserById() {
 
-    @DeleteMapping(path = "/{userId}")
-    public @ResponseBody
-    void deleteUserById(@PathVariable Integer userId) {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        };
+        Integer userId = ((JotUserDetails)principal).getId();
 
-        userRepository.deleteById(userId);
+        Optional<User> user = userRepository.findById(userId);
+
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity(user.get(), HttpStatus.OK);
     }
 
-    @PutMapping(path = "update/{userId}")
-    void updateUser(@PathVariable Integer userId, @RequestParam String firstName, @RequestParam String lastName, @RequestParam String emailAddress,
+    @DeleteMapping(path = "/me")
+    public ResponseEntity deleteUserById() {
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        };
+        Integer userId = ((JotUserDetails)principal).getId();
+
+        Optional<User> user = userRepository.findById(userId);
+
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else {
+            userRepository.deleteById(userId);
+        }
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    }
+
+    @PutMapping(path = "/me")
+    ResponseEntity updateUser(@RequestParam String firstName, @RequestParam String lastName, @RequestParam String emailAddress,
                     @RequestParam String phoneNumber) {
-        User updatedUser = userRepository.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException());
+
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!(principal instanceof JotUserDetails)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        };
+        Integer userId = ((JotUserDetails)principal).getId();
+
+        Optional<User> user = userRepository.findById(userId);
+
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        User updatedUser = user.get();
+
         updatedUser.setFirstName(firstName);
         updatedUser.setLastName(lastName);
         updatedUser.setEmailAddress(emailAddress);
         updatedUser.setPhoneNumber(phoneNumber);
         User savedUser = userRepository.save(updatedUser);
 
+        return new ResponseEntity(savedUser, HttpStatus.OK);
     }
 
 }
